+++ Example/Emission/AppDelegate.m
@@ -27,6 +27,7 @@
 #import <Emission/ARShowMoreInfoComponentViewController.h>
 #import <Emission/ARConversationComponentViewController.h>
 #import <Emission/ARFairMoreInfoComponentViewController.h>
+#import <Emission/ARCityFairListComponentViewController.h>
 #import <Emission/ARCitySectionListComponentViewController.h>
 #import <Emission/ARFairComponentViewController.h>
 #import <Emission/ARFairBoothComponentViewController.h>
@@ -274,6 +275,7 @@ randomBOOL(void)
   UIViewController *viewController = nil;
 
   BOOL isShow = [route hasPrefix:@"/show/"] || [route hasPrefix:@"show/"];
+  BOOL isCityFairList = [route hasPrefix:@"/city-fair/"];
   BOOL isCityList = [route hasPrefix:@"/city/"];
 
   if ([route hasPrefix:@"/artist/"] && [route componentsSeparatedByString:@"/"].count == 3) {
@@ -347,6 +349,11 @@ randomBOOL(void)
   } else if (isShow && [route hasSuffix:@"/info"]) {
     NSString *showID = [[route componentsSeparatedByString:@"/"] objectAtIndex:2];
     viewController = [[ARShowMoreInfoComponentViewController alloc] initWithShowID:showID];
+
+  } else if (isCityFairList) {
+    NSString *citySlug = [[route componentsSeparatedByString:@"/"] lastObject];
+    viewController = [[ARCityFairListComponentViewController alloc] initWithCitySlug:citySlug];
+
   } else if (isCityList) {
     NSString *citySlug = [[route componentsSeparatedByString:@"/"] objectAtIndex:2];
     NSString *section = [[route componentsSeparatedByString:@"/"] objectAtIndex:3];