+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Handle rendering card for Clusters - luc
 - Refactor clustering engine to handle multiple zooms - luc
 - All event section events now link to show - kieran
+- Adds CityFairList - kieran
 
 ### 1.8.11